chore: move benchmarks folder to top level
