+++ packages/store-sync/benchmarks/getRecords.bench.ts
@@ -1,11 +1,11 @@
 import { bench, describe } from "vitest";
 import { getComponentEntities, getComponentValue } from "@latticexyz/recs";
-import { buildTable, getTables } from "../sqlite";
-import { logsToBlocks } from "../../test/logsToBlocks";
-import { createRecsStorage, createSqliteStorage, createZustandStorage, tables } from "../../test/utils";
-import worldRpcLogs10 from "../../../../test-data/world-logs-10.json";
-import worldRpcLogs100 from "../../../../test-data/world-logs-100.json";
-import worldRpcLogs1000 from "../../../../test-data/world-logs-1000.json";
+import { buildTable, getTables } from "../src/sqlite";
+import { logsToBlocks } from "../test/logsToBlocks";
+import { createRecsStorage, createSqliteStorage, createZustandStorage, tables } from "../test/utils";
+import worldRpcLogs10 from "../../../test-data/world-logs-10.json";
+import worldRpcLogs100 from "../../../test-data/world-logs-100.json";
+import worldRpcLogs1000 from "../../../test-data/world-logs-1000.json";
 
 describe.each([
   { numRecords: 10, logs: worldRpcLogs10 },
